Added camera collision and Corrected exit functionality
The camera is treated as a sphere, and the camera's pos and sphere's pos
will be kept the same.

The Game will exit on LEFT CLICKING at the exit screen as required.
Pressing ESC again will resume the program.

Note: Collision can be improved with implementation of a 'clampToWall'
function. (See TODO)